requirements.txt
Flask==1.1.4
Werkzeug==0.16.1
transformers
torch
huggingface-hub
python-dotenv
gradio